+++ src/components/index.ts
@@ -1,7 +1,9 @@
 import Calendar from "./calendar";
+import Event from "./calendar/event";
+
 import CalendarButtons from "./calendarButtons";
 import Header from "./header";
 import Footer from "./footer";
-import SettingModal from "./settingModal";
+import SettingModal from "./modals/settingModal";
 
-export { Calendar, CalendarButtons, Header, Footer, SettingModal };
+export { Calendar, CalendarButtons, Header, Footer, SettingModal, Event };
