+++ src/helpers/element.tsx
@@ -1,4 +1,4 @@
-import CalendarEvent from "@/components/calendarEvent";
+import { Event } from "@/components";
 import styles from "@/styles/components/Calendar.module.scss";
 import dayjs from "dayjs";
 
@@ -9,7 +9,7 @@ const createWeekDaysList = (weekdays: string[]): JSX.Element[] => {
 const createEvents = (data: any[], date: any) => {
   const results: any[] = data[date];
   if (results) {
-    return results.map((an: Activity) => <CalendarEvent activity={an} total={results.length} key={an.anime_id + date + an.status} />);
+    return results.map((an: Activity) => <Event activity={an} total={results.length} key={an.anime_id + date + an.status} />);
   }
 };
 
@@ -30,11 +30,6 @@ const createDaysRow = (data: any[], days: CalendarDay[]): JSX.Element[] => {
       </div>
       <div className={styles.events}>{createEvents(data, day.date)}</div>
     </div>
-
-    // <li key={day.date + day.dayOfMonth} className={!day.isCurrentMonth ? styles.calendar_day__not_current : styles.calendar_day}>
-    //   <span>{day.dayOfMonth}</span>
-    //   {createEvents(data, day.date)}
-    // </li>
   ));
 };
