+++ src/styles/globals.scss
@@ -56,19 +56,3 @@ li {
     color-scheme: dark;
   }
 }
-
-.tooltip {
-  visibility: hidden;
-  width: max-content;
-  position: absolute;
-  top: 0;
-  left: 0;
-  padding: 8px 16px;
-  border-radius: 3px;
-  font-size: 90%;
-  pointer-events: none;
-  opacity: 0;
-  transition: opacity 0.3s ease-out;
-  will-change: opacity, visibility;
-  background-color: aqua;
-}
