+++ src/styles/pages/Home.module.scss
@@ -1,4 +1,4 @@
-@import "./mixins";
+@import "../mixins";
 
 .main {
   justify-content: space-between;
